Extract DaysOfTheWeek as functional component
